Returning full interpolation, using long double internally to prevent nan.

+++ solver.h
@@ -14,7 +14,7 @@ struct rparams {
 
 void print_state (size_t iter, gsl_multiroot_fsolver * s);
 
-double P( double lambda, int b, int i, void * params, double* fas );
+long double P( double lambda, int b, int i, void * params, double* fas );
 
 int equation( const gsl_vector * x, void * params, gsl_vector *eqn );
